Move `SynchronousObserver` into `grakn-grpc`
